add: test renamed and refactorized. #2391
