+++ tests/integration/test_logcollector/test_localfile/data/location_config.yaml
@@ -1,0 +1,12 @@
+---
+- tags:
+  - test_localfile_bad_configuration
+  apply_to_modules:
+  - test_localfile_bad_configuration
+  sections:
+    - section: localfile
+      attributes:
+        - name: 'localfile_bad_config'
+      elements:
+      - log_format:
+          value: LOG_FORMAT
